The execution framework will be stored in the /src/control directory in a series of classes.
The first class is the Driver class. It performs a series of sweeps until a stopping condition is met.

The second class is the Sweep class. This class handles looping through time groupings and proposing updates.

The third class is the Update class. This class handles the logic of randomly selecting accept/reject and insert/delete. It also calls Configuration public member functions to carry out the updates. In the future, the logic for the RANDOM hamiltonian will be moved to the hamiltonians directory along with other hamiltonian models, and Update will just handle the control logic.