NOT TESTED FOR BUILD OR RUN.
Output: This commit contains much of the skeletal infrastructure for standard outputs, data outputs, and RESTART outputs without actually carrying out the details. Output is handled centrally in the Output class to avoid having output code scattered throughout the system.

Sweeps need to store their place in the iteration process (their id) so that caching in Output can be performed.