NOT TESTED FOR BUILD OR RUN.
Random: In the declaration and definition of the Random class, the upper and lower bounds were incorrectly marked as int's, when they should have been doubles. So whenever the width of the groups dipped below 1 K, the bounds would collapse to [0, 0] or similar.

Driver.run: We have to pass the Output object to Driver.run in order to print out RESTART files. Outputs for these files are not yet implemented.