RegionData: Consolidated various time group attributes into a RegionData structure to improve readability and limit the number of arguments functions had to have.
Constant Number of Time Groups: The number of time groups must be constant for detailed balance to be satisfied, so I removed the code that executes time group updates without actually removing the functions themselves.

Sweep.hpp: There was a bug that resulting in a stale itLow iterator that occured when inserting into an empty taus set or inserting with a tau value lower than the lowest value in the set. Solved by Checking std::prev(itLow), and refreshing iterators if either of those cases occurred. Additionally, the if statement in the removal section had a logical error that I addressed.