HamiltonianBase: Almost all of the functionality of the Random model carries over to the tV Hamiltonian (and probability others). So, I created HamiltonianBase to be the parent class to Random and TVModel. HamiltonianBase is not currently functioning in the code, just its basic implementation has been copied from Random.hpp and methods separated into public, private, and protected.
input_example.yaml: After conversations with Dr. Huffman, in 1D we will only have bond lengths of 1 or 2. This is why I have changed bond_type_props.

TVModel: Because TVModel will use almost all of the code in Random, I copied it over to test that the setup was functional. This is temporary, and will be replaced by inheritance from HamiltonianBase.