Eigen3: This is the linear algebra package I am using for now. As my requirements change, I may look into other packages.
Configuration: Because the HSum matrix depends only on the bonds currently in the configuration (plus or minus a single bond), I made its computation a method of the Configuration class.

TVModel: There are several debug statements scattered throughout here right now. The problem is that the weights (Wk, Wkp1, Wkm1) are growing in magnitude and becoming large as the number of sweeps increases. I believe this to be a bug, or at least instability.