Update.cpp: No content in Update.cpp, so removing from the framework.
Configuration.cpp: Removing some outdated TODO statements.

Driver.cpp/hpp: Removing an outdated TODO. Instead of passing the starting sweep number to the Driver constructor, I passed it to the Driver::run() function. I also removed the numTimeGroups data member from Driver because it is not used anywhere in the class.

Output.cpp: Replaced explicit redeclarations of outfile names with references to functions that fulfill the same purpose for simplicity.