Input Scripts: I changed the structure of the example input scripts folder to give more flexibility and specialization. Will be useful now as I transition to 2D lattices and in the future as I need to demonstrate more examples.
Input.hpp: I included y and z parameters in the yaml conversion definition, implemented validation functions for the two and three dimensional inputs, and updated the << operator for the LatticeInput class in order to include information about the other dimensions. The code is very "space conservative", but it is fairly straightforward. I think a bit of repetition in Input code is ok because there will probably be many divergence points from the standard situation in the future. I can always come back and condense this if needed.