LatticeBase: Included LatticeInput input as a stored data member of LatticeBase. I think it eliminates the need for function arguments in child classes, and decreases the need for additional stored data members. This change might be reversed in the future.
SimpleCubic::getNearestNeighborsNEW(): This function returns a vector of nearest neighbors (in no rigorous order for now) for the site you pass in. I tried to eliminate duplicate code for each dimension using a Helper struct and loops, but it may have been simpler to just go through each dimension manually.

Input: The default BC type needs to be OPEN for the following reason. If we are doing a 1D lattice, the getNearestNeighborNEW function of SimpleCubic will return nearest neighbors in the y and z dimensions if those BCs were periodic. I implemented default BC type definitions and validation checks to require this.