CHECKPOINT COMMIT!!!!!
It is clear that creating support for the new Site structs within Lattices, Bonds, Configurations and Hamiltonians will be too complex if I try to keep the existing implementations as I do it. So, this is the final commit where both old and new versions will be compatible. From this point on I will remove old implementations and replace them with the new ones. This commit is the one to return to if things go sideways.

Bond: I have chosen to store the sites for a bond in a set of pointers to Site objects. This eliminates the need to copy the data from Site objects into new sites, instead just referencing the existing objects stored in SimpleCubic::sitesNEW. The pointers are labelled "const" so that the Site objects cannot be changed by calling functions. A forward declaration of Site was required in Bond.hpp (and will probably be required elsewhere) for compilation.

Configuration: Due to ambiguity between the two constructors for Bond currently implemented, I needed to explicitly define the type of the empty set for nullBond.