NON-FUNCTIONING: This commit is non-functioning. Committing due to a lot of compounding changes and in an effort to avoid progress loss.
LatticeHolder: This static inline function allows a lattice pointer to be obtained from anywhere in the program. Useful because the lattice never needs to be updated after being created, and I need the lattice in the Configuration::operator>> definition.

Bond: The new format for reading in bonds has the six cartesian coordinates of the first and second site listed out on one line. It is simple and easy to incorporate with the existing Configuration::operator>>. The sites included in a bond are constant pointers to Site objects in the sites vector of a lattice. SiteSumLess struct is necessary to define an ordering method for the set of Site objects.

Lattices:
The Site structs are fundamentally the same for Honeycomb and SimpleCubic. So, I am defining a SiteBase struct that the two Site structs inherit from. This allows for further customization for each lattice type in the future while also having a shared structure.