NON-FUNCTIONING: This commit is non-functioning. Committing due to accumulation of high impact changes.
RandomHelpers: The uniform probability functions with bounds needed checks that upper bounds were greater than lower bounds.

SimpleCubic: The selection of a random site at which to start a bond seems to fit better within the Lattice framework, because it has all of the information about BC's, different dimensions, etc that is needed to choose a random site. Getting nearest neighbors of a site is alkso a Lattice centered task.

HamiltonianBase: I began creating support for the new Bond/Site style by addressing the createBondToInsert function.