CHECKPOINT COMMIT!!!!! This commit has been tested for runtime errors in a variety of situations (1D, 2D, 3D, combinations of OPEN and PERIODIC BC's). Behavior appears to qualitatively match behavior before the multidimensional changes (previous CHECKPOINT COMMIT).
RandomHelpers/SimpleCubic::chooseStartInd: In RandomHelpers, it is possible for low == high if we are trying to get a site index for a dimension that has a thickness of 1 (and thus open BCs. Prior to this change, the behavior was undefined. The current solution is to throw an error if low == high, and instead handle the thickness of 1 situation in SimpleCubic::chooseStartInd with a simple if statement. This leaves lattice specific logic out of the general RandomHelper functions.

LatticeBase: I am now including bondLength as an input to getNumUniqueBonds and chooseRandSite in case we want to look at length 1 bonds, or potential >2 site interactions in the future. LatticeBase has the default value of bondLength = 2 defined in its function signatures.

SimpleCubic::getNumUniqueBonds: Previous to the present changes to this function, the subtractions from nBonds based on NSites in each dimension would occur regardless of what dimensionality that problem had. This would cause nBonds to go to 0 in some cases. These changes avoid this problem by allowing subtractions to occur based on what dimensionality the problem has. In order for this logic to be valid, a 1D problem must be in the x direction, and a 2D problem must be in the x and y direction. This still needs to be enforced in Input.hpp.